requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
slowapi==0.1.8

# Google Gemini Integration
google-generativeai==0.3.2
google-cloud-aiplatform==1.38.1

# Neo4j Integration
neo4j==5.15.0
py2neo==2021.2.4

# e2b.dev Integration
e2b==0.15.0

# Data Processing & Analytics
pandas==2.1.4
numpy==1.25.2
scipy==1.11.4
scikit-learn==1.3.2
networkx==3.2.1

# Graph Analytics & ML
torch==2.1.1
torch-geometric==2.4.0
dgl==1.1.3

# Visualization
matplotlib==3.8.2
plotly==5.17.0
seaborn==0.13.0

# Image Processing & OCR
Pillow==10.1.0
opencv-python==4.8.1.78
pytesseract==0.3.10

# Web & API
httpx==0.25.2
aiohttp==3.9.1
websockets==12.0

# Database & ORM
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0

# Security & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Configuration & Environment
python-dotenv==1.0.0

# Logging & Monitoring
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0
pythonjsonlogger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development Tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Blockchain & Crypto (for crypto fraud detection)
web3==6.12.0
eth-account==0.9.0

# Financial Data APIs
yfinance==0.2.28
alpha-vantage==2.3.1

# Natural Language Processing
spacy==3.7.2
transformers==4.36.2

# Time Series Analysis
statsmodels==0.14.0

# Async & Concurrency
asyncio==3.4.3
aiofiles==23.2.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3
click==8.1.7
rich==13.7.0
tqdm==4.66.1
